refactor: use original sqlalchemy

+++ src/main.py
@@ -14,7 +14,7 @@ from fastapi.responses import ORJSONResponse
 from pydantic import BaseModel, Field
 from rich import get_console
 from rich.progress import track
-from sqlalchemy import func
+from sqlalchemy import Select, func, select
 from sqlalchemy.orm import Session, joinedload
 from starlette.convertors import Convertor, register_url_convertor
 from starlette.middleware.gzip import GZipMiddleware
@@ -22,7 +22,18 @@ from wdtagger import Tagger
 
 import shared
 from ai import OpenAIImageAnnotator
-from models import Post, PostBase, PostHasTag, PostWithTag, Tag, TagGroup, TagPublic
+from models import (
+    Post,
+    PostHasTag,
+    PostPublic,
+    PostWithTagPublic,
+    Tag,
+    TagGroup,
+    TagGroupPublic,
+    TagGroupWithTagsPublic,
+    TagPublic,
+    TagWithGroupPublic,
+)
 from utils import (
     attach_tags_to_post,
     delete_by_file_path_and_ext,
@@ -68,12 +79,9 @@ class PathConvertor(Convertor):
 register_url_convertor("pathlike", PathConvertor())
 
 
-def get_post_by_id(post_id, session) -> PostWithTag:
-    post = (
-        session.query(Post)
-        .filter(Post.id == post_id)
-        .options(joinedload(Post.tags).joinedload(PostHasTag.tag_info))
-        .first()
+def get_post_by_id(post_id, session) -> PostWithTagPublic:
+    post = session.get(
+        Post, post_id, options=[joinedload(Post.tags).joinedload(PostHasTag.tag_info).joinedload(Tag.group)]
     )
     if post is None:
         raise HTTPException(status_code=404, detail="Post not found")
@@ -92,17 +100,18 @@ class PostFilter(BaseModel):
 
 @app.post(
     "/v1/posts",
-    response_model=list[PostBase],
+    response_model=list[PostPublic],
     tags=["Post"],
 )
-def v1_get_posts(
+def v1_list_posts(
     limit: int | None = None,
     offset: int = 0,
     filter: PostFilter = PostFilter(),
 ):
+
     session = get_session()
-    query = apply_filtered_query(filter, session.query(Post)).limit(limit).offset(offset)
-    return query.all()
+    stmt = apply_filtered_query(filter, select(Post)).limit(limit).offset(offset)
+    return session.scalars(stmt).unique().all()
 
 
 @app.delete("/v1/posts/{post_id}", tags=["Post"])
@@ -116,18 +125,18 @@ def v1_delete_post(post_id: int):
     return
 
 
-def apply_filtered_query(filter: PostFilter, query: fastapi.Query):
+def apply_filtered_query(filter: PostFilter, stmt: Select) -> Select:
     if filter.rating:
-        query = query.filter(Post.rating.in_(filter.rating))
+        stmt = stmt.filter(Post.rating.in_(filter.rating))
     if filter.score:
-        query = query.filter(Post.score.in_(filter.score))
+        stmt = stmt.filter(Post.score.in_(filter.score))
     if filter.tags:
-        query = query.join(Post.tags).filter(PostHasTag.tag_name.in_(filter.tags))
+        stmt = stmt.join(Post.tags).filter(PostHasTag.tag_name.in_(filter.tags))
     if filter.extension:
-        query = query.filter(Post.extension.in_(filter.extension))
+        stmt = stmt.filter(Post.extension.in_(filter.extension))
     if filter.folder and filter.folder != "":
-        query = query.filter(Post.file_path == filter.folder)
-    return query
+        stmt = stmt.filter(Post.file_path == filter.folder)
+    return stmt
 
 
 class PostCountResponse(BaseModel):
@@ -149,12 +158,12 @@ class RatingCountResponse(BaseModel):
 @app.post("/v1/posts/count/rating", response_model=list[RatingCountResponse], tags=["Post"])
 def v1_count_group_by_rating(
     filter: PostFilter = Body(...),
+    session: Session = Depends(get_session),
 ):
 
-    session = get_session()
-    query = session.query(Post.rating, func.count()).group_by(Post.rating)
-    query = apply_filtered_query(filter, query)
-    resp = query.all()
+    stmt = select(Post.rating, func.count()).group_by(Post.rating)
+    stmt = apply_filtered_query(filter, stmt)
+    resp = session.execute(stmt).all()
     return [RatingCountResponse(rating=row[0] if row[0] is not None else 0, count=row[1]) for row in resp]
 
 
@@ -166,11 +175,10 @@ class ScoreCountResponse(BaseModel):
 @app.post("/v1/posts/count/score", response_model=list[ScoreCountResponse], tags=["Post"])
 def v1_count_group_by_score(
     filter: PostFilter = Body(...),
+    session: Session = Depends(get_session),
 ):
-    session = get_session()
-    query = session.query(Post.score, func.count()).group_by(Post.score)
-    query = apply_filtered_query(filter, query)
-    resp = query.all()
+    query = apply_filtered_query(filter, select(Post.score, func.count()).group_by(Post.score))
+    resp = session.execute(query).all()
     return [ScoreCountResponse(score=row[0] if row[0] is not None else 0, count=row[1]) for row in resp]
 
 
@@ -182,11 +190,11 @@ class ExtensionCountResponse(BaseModel):
 @app.post("/v1/posts/count/extension", response_model=list[ExtensionCountResponse], tags=["Post"])
 def v1_count_group_by_extension(
     filter: PostFilter = Body(...),
+    session: Session = Depends(get_session),
 ):
-    session = get_session()
-    query = session.query(Post.extension, func.count()).group_by(Post.extension)
+    query = select(Post.extension, func.count()).group_by(Post.extension)
     query = apply_filtered_query(filter, query)
-    resp = query.all()
+    resp = session.execute(query).all()
     return [ExtensionCountResponse(extension=row[0], count=row[1]) for row in resp]
 
 
@@ -194,7 +202,7 @@ class ScoreUpdate(BaseModel):
     score: Annotated[int, Field(ge=0, le=5)]
 
 
-@app.put("/v1/posts/{post_id}/score", response_model=Post, tags=["Post"])
+@app.put("/v1/posts/{post_id}/score", response_model=PostWithTagPublic, tags=["Post"])
 def v1_update_post_score(post_id: Annotated[int, Path(gt=0)], score_update: ScoreUpdate):
     session = get_session()
     post = session.query(Post).filter(Post.id == post_id).first()
@@ -209,40 +217,45 @@ class RatingUpdate(BaseModel):
     rating: Annotated[int, Field(ge=0, le=5)]
 
 
-@app.put("/v1/posts/{post_id}/rating", response_model=Post, tags=["Post"])
-def v1_update_post_rating(post_id: Annotated[int, Path(gt=0)], rating_update: RatingUpdate):
-    session = get_session()
+@app.put("/v1/posts/{post_id}/rating", response_model=PostPublic, tags=["Post"])
+def v1_update_post_rating(
+    post_id: Annotated[int, Path(gt=0)],
+    rating_update: RatingUpdate,
+    session: Session = Depends(get_session),
+):
     post = session.query(Post).filter(Post.id == post_id).first()
     if not post:
         raise HTTPException(status_code=404, detail="Post not found")
     post.rating = rating_update.rating
     session.commit()
+    session.refresh(post)
     return post
 
 
-@app.put("/v1/posts/{post_id}/source", response_model=Post, tags=["Post"])
-def v1_update_post_source(post_id: Annotated[int, Path(gt=0)], source: str):
-    session = get_session()
-    post = session.query(Post).filter(Post.id == post_id).first()
+@app.put("/v1/posts/{post_id}/source", response_model=PostPublic, tags=["Post"])
+def v1_update_post_source(post_id: Annotated[int, Path(gt=0)], source: str, session: Session = Depends(get_session)):
+    post = session.get(Post, post_id)
     if not post:
         raise HTTPException(status_code=404, detail="Post not found")
     post.source = source
     session.commit()
+    session.refresh(post)
     return post
 
 
-@app.put("/v1/posts/{post_id}/caption", response_model=Post, tags=["Post"])
+@app.put("/v1/posts/{post_id}/caption", response_model=PostPublic, tags=["Post"])
 def v1_update_post_caption(post_id: Annotated[int, Path(gt=0)], caption: str):
     session = get_session()
-    post = session.query(Post).filter(Post.id == post_id).first()
+    post = session.get(Post, post_id)
     if not post:
         raise HTTPException(status_code=404, detail="Post not found")
     post.caption = caption
     session.commit()
+    session.refresh(post)
     return post
 
 
-@app.get("/v1/posts/{post_id}", response_model=PostWithTag, tags=["Post"])
+@app.get("/v1/posts/{post_id}", response_model=PostWithTagPublic, tags=["Post"])
 def v1_get_post(post_id: int):
     session = get_session()
     return get_post_by_id(post_id, session)
@@ -250,17 +263,21 @@ def v1_get_post(post_id: int):
 
 @app.get("/v1/images/{post_path:pathlike}", tags=["Image"])
 def v1_get_post_by_path(post_path: str):
+    if not shared.target_dir:
+        raise HTTPException(status_code=400, detail="Target directory is not set")
     abs_path = shared.target_dir / post_path
     return fastapi.responses.FileResponse(abs_path)
 
 
 @app.get("/v1/thumbnails/{post_path:pathlike}", tags=["Image"])
 def v1_get_thumbnail(post_path: str):
+    if not shared.thumbnails_dir:
+        raise HTTPException(status_code=400, detail="Thumbnails directory is not set")
     abs_path = shared.thumbnails_dir / post_path
     return fastapi.responses.FileResponse(abs_path)
 
 
-@app.put("/v1/cmd/posts/{post_id}/rotate", response_model=Post, tags=["Command"])
+@app.put("/v1/cmd/posts/{post_id}/rotate", response_model=PostWithTagPublic, tags=["Command"])
 def v1_cmd_rotate_image(post_id: int, clockwise: bool = True, session: Session = Depends(get_session)):
     post = session.get(Post, post_id)
     if post is None:
@@ -269,25 +286,29 @@ def v1_cmd_rotate_image(post_id: int, clockwise: bool = True, session: Session =
     return post
 
 
+class TagAndGroupIdPublic(BaseModel):
+    name: str
+    group_id: int
+
+
 class TagResponse(BaseModel):
     count: int
-    tag_info: TagPublic
+    tag_info: TagAndGroupIdPublic
 
 
 @app.get("/v1/tags", response_model=list[TagResponse], tags=["Tag"])
 def v1_get_tags():
     session = get_session()
-    # 从 PostHasTag 表中查询所有的 tag_name 和 count，从 Tag 表中查询 tag_name 对应的 Tag 实例
-    query = (
-        session.query(PostHasTag.tag_name, func.count(), Tag)
-        .group_by(PostHasTag.tag_name)
-        .join(Tag, PostHasTag.tag_name == Tag.name)
-    )
-    resp = query.all()
+    resp = session.execute(
+        select(PostHasTag.tag_name, func.count(), Tag.group_id)
+        .select_from(PostHasTag)
+        .join(Tag)
+        .group_by(PostHasTag.tag_name),
+    ).all()
     return [
         TagResponse(
             count=row[1],
-            tag_info=TagPublic(name=row[0], group_id=row[2].group_id),
+            tag_info=TagAndGroupIdPublic(name=row[0], group_id=row[2]),
         )
         for row in resp
     ]
@@ -321,15 +342,19 @@ def v1_get_tag(tag_name: str):
 def v1_update_tag(tag_name: str, new_tag_name: str):
     session = get_session()
     tag = session.query(Tag).filter(Tag.name == tag_name).first()
+    if tag is None:
+        raise HTTPException(status_code=404, detail="Tag not found")
     tag.name = new_tag_name
     session.commit()
     return tag
 
 
-@app.post("/v1/posts/{post_id}/tags/{tag_name}", response_model=PostWithTag, tags=["Tag"])
+@app.post("/v1/posts/{post_id}/tags/{tag_name}", response_model=PostWithTagPublic, tags=["Tag"])
 def v1_add_tag_to_post(post_id: int, tag_name: str):
     session = get_session()
     post = session.query(Post).filter(Post.id == post_id).first()
+    if post is None:
+        raise HTTPException(status_code=404, detail="Post not found")
     tag = session.query(Tag).filter(Tag.name == tag_name).first()
     # 如果 tag 不存在，创建一个新的 tag
     if tag is None:
@@ -341,18 +366,20 @@ def v1_add_tag_to_post(post_id: int, tag_name: str):
         session.query(PostHasTag).filter(PostHasTag.post_id == post_id, PostHasTag.tag_name == tag_name).first()
     )
     if post_has_tag is None:
-        postHasTag = PostHasTag(post_id=post_id, tag_name=tag_name, is_auto=False)
+        postHasTag = PostHasTag(post=post, post_id=post_id, tag_name=tag_name, tag_info=tag)
         session.add(postHasTag)
         session.commit()
-    post = get_post_by_id(post_id, session)
-    return post
+    return get_post_by_id(post_id, session)
 
 
-@app.delete("/v1/posts/{post_id}/tags/{tag_name}", response_model=PostWithTag, tags=["Tag"])
+@app.delete("/v1/posts/{post_id}/tags/{tag_name}", response_model=PostWithTagPublic, tags=["Tag"])
 def v1_remove_tag_from_post(post_id: int, tag_name: str):
     session = get_session()
-    post = session.query(Post).filter(Post.id == post_id).first()
-    tag_record = session.query(PostHasTag).filter(PostHasTag.tag_name == tag_name).first()
+    post = session.get(Post, post_id)
+    if post is None:
+        raise HTTPException(status_code=404, detail="Post not found")
+    # tag_record = session.query(PostHasTag).filter(PostHasTag.tag_name == tag_name).first()
+    tag_record = session.get(PostHasTag, (post_id, tag_name))
     if tag_record is not None:
         session.delete(tag_record)
         session.commit()
@@ -360,7 +387,7 @@ def v1_remove_tag_from_post(post_id: int, tag_name: str):
     return post
 
 
-@app.get("/v1/tag-groups", response_model=list[TagGroup], tags=["Tag"])
+@app.get("/v1/tag-groups", response_model=list[TagGroupWithTagsPublic], tags=["Tag"])
 def v1_get_tag_groups():
     session = get_session()
     return session.query(TagGroup).all()
@@ -375,12 +402,13 @@ def v1_cmd_process_posts():
 @app.get("/v1/cmd/auto-tags/{post_id}", response_model=Post, tags=["Command"])
 def v1_cmd_auto_tags(post_id: int):
     session = get_session()
-    post = session.query(Post).filter(Post.id == post_id).first()
-
+    # post = session.query(Post).filter(Post.id == post_id).first()
+    post = session.get(Post, post_id)
+    if post is None:
+        raise HTTPException(status_code=404, detail="Post not found")
     abs_path = post.absolute_path
-    if shared.tagger is None:
-        shared.tagger = Tagger(model_repo="SmilingWolf/wd-vit-large-tagger-v3", slient=True)
-    resp = shared.tagger.tag(abs_path)
+    tagger = shared.tagger if shared.tagger is not None else Tagger(model_repo="SmilingWolf/wd-vit-large-tagger-v3")
+    resp = tagger.tag(abs_path)
     shared.logger.info(resp)
     post.rating = from_rating_to_int(resp.rating)
     attach_tags_to_post(session, post, resp, is_auto=True)
@@ -511,7 +539,7 @@ def v1_upload_file(
         headers = {
             "user-agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/127.0.0.0 Safari/537.36",
         }
-        if "pximg.net" in url:
+        if url and "pximg.net" in url:
             headers["referer"] = "https://www.pixiv.net/"
         with httpx.Client() as client:
             resp = client.get(url, headers=headers)
@@ -520,10 +548,10 @@ def v1_upload_file(
     else:
         file_io = file.file
 
-    if path is None and file is not None:
+    if path is None and file is not None and file.filename:
         path = file.filename
     else:
-        path = path or url.split("/")[-1]
+        path = path or (url.split("/")[-1] if url else "")
 
     abs_path = shared.target_dir / path
     abs_path.parent.mkdir(parents=True, exist_ok=True)
